+++ app/models/rails_workflow/operation.rb
@@ -10,6 +10,7 @@ module RailsWorkflow
     include Operations::Assignments
 
     belongs_to :process, class_name: 'RailsWorkflow::Process'
+    alias_method :parent, :process
     belongs_to :template, class_name: 'RailsWorkflow::OperationTemplate'
     belongs_to :child_process, class_name: 'RailsWorkflow::Process'
     has_one :context, class_name: 'RailsWorkflow::Context', as: :parent